# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurante API",
        "description": "Documentación del backend (Laravel + Swagger).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Listar usuarios (paginado)",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Usuario"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "usuario",
                                    "password",
                                    "nombre",
                                    "correo",
                                    "rol"
                                ],
                                "properties": {
                                    "usuario": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secreto123"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Demo",
                                        "nullable": true
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "rol": {
                                        "description": "admin|gerente|cajero|mozo|cocinero|empleado",
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Ver usuario",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar usuario",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "usuario": {
                                        "type": "string",
                                        "example": "admin2"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "nuevo123"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Admin"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Demo",
                                        "nullable": true
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin2@example.com"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "example": "gerente"
                                    },
                                    "activo": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "restaurant_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Usuario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Usuario": {
                "required": [
                    "usuario",
                    "password",
                    "nombre",
                    "correo"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "usuario": {
                        "type": "string",
                        "example": "admin"
                    },
                    "password": {
                        "type": "string",
                        "example": "secreto123"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Demo",
                        "nullable": true
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "admin@example.com"
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}